+++ modules/system/lang/en/validation.php
@@ -46,6 +46,7 @@ return array(
         "array"   => "The :attribute may not have more than :max items.",
     ),
     "mimes"            => "The :attribute must be a file of type: :values.",
+    "extensions"       => "The :attribute must have an extension of: :values.",
     "min"              => array(
         "numeric" => "The :attribute must be at least :min.",
         "file"    => "The :attribute must be at least :min kilobytes.",